+++ History.md
@@ -3,6 +3,7 @@
 ### Version 4.10.1 [not published yet]
 
 - Public votings assigned to user groups now also show the users that have not voted yet in the admin view.
+- The voting result page now has a full screen mode and does not reload automatically anymore.
 - Bugfix: Some constellations of whitespaces and line breaks at the end of list points could lead to problems with amendments.
 - Bugfix: When saving binary files to the file system was enabled, copying / moving motions to other consultations broke uploaded images and PDFs.